delete prompt is implemented
anand

+++ v-admin/v-includes/functions/function.editGalleryThumb.php
@@ -25,12 +25,19 @@
 	
 	//upload the selected thumb image in the templ folder using 
 	//file uploader class in the temp folder
-	$filename_temp = $uploadFile->upload_file($galleryId,'gallery_thumb',$outputPath_temp);
+	echo $filename_temp = $uploadFile->upload_file($galleryId,'gallery_thumb',$outputPath_temp);
 	
-	//convert the file to get the .JPG extension
-	$mediaQuery->convertImageFormat($outputPath_temp.$filename_temp,$outputPath_temp.$galleryId,".JPG");
-	//delete the previous file
-	unlink($outputPath_temp.$filename_temp);
+	$file_info = pathinfo($outputPath_temp.$filename_temp);
+	$ext = $file_info['extension'] ;
+	
+	if( $ext != "JPG")
+	{
+		//convert the file to get the .JPG extension
+		$mediaQuery->convertImageFormat($outputPath_temp.$filename_temp,$outputPath_temp.$galleryId,".JPG");
+		
+		//delete the previous file
+		unlink($outputPath_temp.$filename_temp);
+	}
 	
 	//Open the file for croping in crop page
